4.15-4.23

+++ osa4/bloglist/controllers/blogs.js
@@ -1,26 +1,35 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const userExtractor = require('../utils/middleware').userExtractor
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const { title, author, url, likes } = request.body
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  const user = request.user
+  const body = request.body
 
-  if (!title || !url) {
+  if (!body.title || !body.url) {
     return response.status(400).json({ error: 'title or url missing' })
   }
 
   const blog = new Blog({
-    title,
-    author,
-    url,
-    likes
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id
   })
 
   const savedBlog = await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
@@ -43,8 +52,22 @@ blogsRouter.put('/:id', async (request, response) => {
   response.json(result)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).end()
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(403).json({
+      error: 'only creator can delete'
+    })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
+
   response.status(204).end()
 })
